backend/requirements.txt
django
djangorestframework
PyJWT
django-cors-headers